remove squeezellm, restructure to clean-up files
